cl/replication: moved sink and source implementation
Moved source and sync implementation out of replication module. This way
the `replication` sub-module does not have to depend on anything from
cluster_link module and have no information about the specific
implementation of actual source and sink that are used in cluster
linking.
