+++ src/v/cluster_link/BUILD
@@ -205,10 +205,11 @@ redpanda_cc_library(
         "//src/v/base",
         "//src/v/cluster",
         "//src/v/cluster/utils:partition_change_notifier_api",
-        "//src/v/cluster_link/replication:deps_impl",
+        "//src/v/cluster_link/replication:deps",
         "//src/v/cluster_link/replication:mux_remote_consumer",
         "//src/v/kafka/client/direct_consumer",
         "//src/v/kafka/server",
+        "//src/v/kafka/server:write_at_offset_stm",
         "//src/v/model",
         "//src/v/raft",
         "//src/v/rpc",
